Refactoring schematic generation.
This commit prepares for the netlist_map dependency of cellview_info.
Basically, if we want to netlist a custom block outside of BAG, we
must have access to the cellview_info objects of the custom blocks
in order to figure out the correct internal net attributes.  This
means we need access to netlist_map (which maps from cellview names
to cellview_info objects) when computing the cellview_info object.

1. change term_attrs member of cellview_info to term_net_attrs, which
   is more accurate.

2. change oa_write.cpp to not use cellview_info class and not use
   boost::range::join, as cellview_info class is going to get more
   complex soon with netlist_map dependency.